+++ p2p/types/router.go
@@ -3,6 +3,8 @@ package types
 //go:generate mockgen -package=mock_types -destination=./mocks/network_mock.go github.com/pokt-network/pocket/p2p/types Router,RouterConfig
 
 import (
+	"context"
+
 	cryptoPocket "github.com/pokt-network/pocket/shared/crypto"
 	"github.com/pokt-network/pocket/shared/modules"
 )
@@ -12,8 +14,10 @@ import (
 type Router interface {
 	modules.IntegratableModule
 
+	Bootstrap(ctx context.Context, maxConcurrency uint32) error
 	Broadcast(data []byte) error
 	Send(data []byte, address cryptoPocket.Address) error
+	Close() error
 
 	// Address book helpers
 	// TECHDEBT: simplify - remove `GetPeerstore`
@@ -22,10 +26,11 @@ type Router interface {
 	RemovePeer(peer Peer) error
 }
 
-type RouterHandler func(data []byte) error
+type MessageHandler func(data []byte) error
 
 // RouterConfig is used to configure `Router` implementations and to test a
 // given configuration's validity.
 type RouterConfig interface {
 	IsValid() error
+	//GetHost() libp2pHost.Host
 }